openssl: prefix libs with 'openssl-'
When people search for ssl in our SBOMs they're typically meaning to search for
libssl and libcrypt which are not showing up in the search.  It's not immediately
obvious that libssl and libcrypt are from openssl or that someone should search
that when they're just looking for information about openssl. This prefixes the
library packages with openssl so that they're more easily discovered.