protobuf: stop pulling in vendored deps
We pretty much never want to update submodules to pull in depencies since we package most dependent libraries.

LIBUPB was turned off in https://github.com/wolfi-dev/os/pull/30312

I am not sure what the actual problem was there's no log saved that I can see it was in service of and image that has not been built. The packages that were conflicting with it are also probably broken.